+++ apps/admin/api/collection/index.ts
@@ -46,3 +46,34 @@ export const useGetCollections = (
     queryKey: [QUERY_KEYS.COLLECTIONS,JSON.stringify(query)],
     queryFn: () => getCollections(query),
   })
+
+
+
+
+/**
+ * Feature collection
+ *
+ * @throws {Error} - Throws an error if there's a problem with the API response.
+ *
+ *  @returns {Promise<object>} - The collection data.
+ */
+
+const featureCollection = async (id: string) => {
+  try {
+    await fetchClient<Collection>(`/v1/collections/${id}/feature`, {
+      method: 'PATCH',
+    })
+  } catch (error) {
+    if (error instanceof Error) {
+      throw error
+    }
+
+    // Error from server.
+    if (error instanceof Response) {
+      const response = await error.json()
+      throw new Error(response.message)
+    }
+  }
+}
+
+export const useFeatureCollection = () => useMutation({mutationFn: featureCollection})
